// findx/JavaScripts/const/Enum.ts
export enum OnClickType {
    /**坐 */
    Sit = 1,
    /**震荡 */
    Shake = 2,
    /**跳舞 */
    Dance = 3,
}

export enum AchievementType {
    /**1-获取宝箱 */
    achType1 = 1,
    /**2-提升等级 */
    achType2 = 2,
    /**3-提升攻击力 */
    achType3 = 3,
    /**4 - 提升血量 */
    achType4 = 4,
    /**5 - 击杀人数 */
    achType5 = 5,
    /**6 - 最大高度 */
    achType6 = 6,
    /**7 - 获取宠物数 */
    achType7 = 7,
    /**8 - 击飞次数 */
    achType8 = 8,
    /**9 - 发射次数 */
    achType9 = 9,
    /**10 - 跳舞次数 */
    achType10 = 10,
    /**11 - 坐下次数 */
    achType11 = 11,
    /**12 - 使用武器次数 */
    achType12 = 12,
    /**13 - 换装次数 */
    achType13 = 13,
    /**14 - 使用翅膀次数 */
    achType14 = 14,
    /**15 - 使用拖尾次数 */
    achType15 = 15,
    /**16 - 切换背景音乐次数 */
    achType16 = 16,
    /**17 - 抽奖次数 */
    achType17 = 17,
    /**18 - 使用藏宝图次数 */
    achType18 = 18,
    achType20 = 20,
    achType21 = 21,
    achType22 = 22,
    achType23 = 23,
    achType24 = 24,
    achType25 = 25,
    achType26 = 26,
    achType27 = 27,
    achType28 = 28,
    achType29 = 29,
}